feat: set up rainbow-delimiters-base-error-face
The default face was nowhere to be seen using the gruvbox-dark
theme so I changed it up to the strongest, more bright red that
I could find, in order to make sure that the error parentheses
are found and corrected.

I used the alexm/set-faces-by-spec function, which is the easiest
way to configure faces in your configuration.