feat: indent and follow symlinks always
As part of this =dot_files= repo, my configuration philosofy uses
symlinks to control the files within the repo but still benefit from
the modifying on only individual files, instead of copying and pasting.
The =vc= emacs package always asks whether I intend to follow the links
that are git-controlled, obviously yes.