fix: remove updated font and hide-leading-stars
I found out that I'm way more of a purist that I expected.
I tried out the configuration with different fonts and headline
stars for org-mode, but I wasn't really comfortable. I feel just
hiding the leading stars and changing the size of the fonts for
the different headlines is enough for me.