fix: update org-agenda view order
I put the upcoming information in the bottom of the page (which gives
the very easy command C-v to check the upcoming items).

I also changed the which-key-idle-delay for it to pop
faster, I found that having it at 1.5 gave me too much idle time,
waiting for the popup to happen.