feat: add C# programming configuration and magit
I changed some magit configuration removing the ^M chars
at the end of each line in the magit-diff buffers.

This is specially useful when changing projects from Windows
to Linux (as is the case with some of my developer friends using
Windows).

I also added csharp-mode and the C# lsp-mode configuration to support
.NET programming.