feat: add linking headings configuration
The Org mode manual section about internal links does refer to
:CUSTOM_ID: properties for internal links. This provides you the
ability to link to a heading, which then gets exported in HTML,
giving the ability to create table of contents and more for
org-mode files. The problem with this is that creating the org-mode
headings ids is really cumbersome. @karl-voit created a few custom
functions which I find interesting and thought including them in my
configuration would improve my linking workflow.