feat: add corfu configuration
As I mention in the description in the Corfu section, this change
is still with the intent of making my Emacs configuration more modular
and easy to read.

I find that Corfu is the perfect combination of having performant
completion candidates, while keeping the buffer out of interruptions
if it is not really need it (this is called idle-completion).

I also added the Default bindings section, on which I intend to change
the bindings that feel too obscure to really be used.