fix: remove lsp-completion-provider
lsp uses company-capf by default for some reason, if
the lsp-completion-provider is set to :capf.

The solution (for it to use the default capf - completion-at
point-function) is to set the lsp-completion-povider to :none,
allowing corfu to take over the completion without any warnings,
as it uses the default completion candidates, maybe given by
lsp, pretty cool.