feat: change emacs config structure adopting stow
I intend to start managing my dotfiles using stow, and there is no
better way than starting with my precious Emacs, which needs its
special configuration files.

I created the stow_directory "emacs", which should be in the structure
that it needs to be to meet my reproducible configuration. A special
note could be that if you are not used to using mu4e, or you haven't
configured it, you should probably untangle the configuration.