refactor: add straight to .gitignore, readme.org
This commit is only a small refactor after configuring
stow to symlink my entire dot_files. I'm pretty proud
of the result, I look forward having the same information
in the different systems I use.

I will document the process in the README further, but this
is all I have time for today.