fix: add custom corfu-global-mode variable
Instead of calling the 'function' corfu-global-mode
in the :config section, I also added the change in
custom variable corfu-global-mode to t.

I'm still trying to get corfu to work perfectly, and
I think I did it both by configuring corfu-global-mode
correctly and the deletion of the indent-line-function,
leaving the configuration to corfu.