fix: make affe work
Contrary to my senses, I thought affe worked —or at
least could be configured— to use the projectile-root
directory to run its queries.

As affe has only been defined to work in the current directory (which
could be enormous by the way) and file paths for the resulting files,
even from the correct directory where called from the current directory
with the path appended.

Changing the consult project file to ignore projectile's projectile-root
and use the default (current directory) fixed it.