feat: add plantuml support to Emacs
There is the great plantuml-mode for Emacs, which provides commands of
the likes of plantuml-preview, which is great for testing the sintax and
then saving the images if you are really interested.

It also provides different execution and saving modes, including server,
executable and jar (the default). The saving modes include: svg, png and
txt if none of the above are installed.

For installing plantuml, I ran "brew install plantuml", which links to
the /usr/local/bin/plantuml binary. Setting up the execution mode to
'executable worked out of the box.