feat: add god-mode to avoid modifier pain
I added god-mode, which is basically a local mode to be enabled,
removing the use of modifiers for the comments. It prepends C- to all
normal keypresses. You also have the g key, which changes to M-.

Using the G key, you can run commands with C-M-. The mode has an easy
adoption process if you are used to your keybindings in normal emacs, as
you just need to NOT press the modifiers.