feat: add general.el and create qk/leader-key-def
This is the first step to start migrating to using Spacemacs bindings in
my config. I find that having modal editing + vim keys will help in the
future, both with remote machines and my own RSI.

That said, I find that using the SPC key as a leader key for most things
is good, specially with buffers and extras.