feat: update keybindings to its own file
Create the file keybindings.el, as the creation of the definers and some
key-pockets was already extensive and hard to read in the literate
config. I plan on doing this for other long source blocks, like org-mode
configuration.

Update the keybindings in the org-agenda-map using evil-mode, to access
schedule easily, which for some reason isn't bound in evil-org.

Create minor-mode-definer and update major-mode-definer for org-mode
files, to include easy access to org-deadline, org-export-dispatch,
org-schedule and org-refile.

Remove unused transients, like the qk/org-agenda-transient. No longer
need to access easy bindings for the actions, as they are already
defined in evil-org.

Finally, delete unused lisp/ files, which were stale for months.