feat: add org-agenda-mode motion and relative line numbers
Configure relative line numbers, which seem to be more interesting and
"force" me to remember about prefixes when doing the evil-mode
commands. It is key to disable it locally for pdf-view-mode, which
breaks totally if the line numbers are enabled.

Set the initial evil state to 'motion for org-agenda, to avoid possible
flashes to Emacs mode (<E> in the modeline) the first the org-agenda is opened.