feat: change tree-sitter-mode to config.org
I think tree-sitter-mode should be in the default config, as it is way
better for all supported programming nodes, faster parsing and opening
of files.

I'm thinking of swapping lsp-mode which is too big for my taste to only
use eglot, which is supposed to be smaller and more focused.