feat: clean up lsp-mode configuration
I decided to invest time cleaning up lsp-mode instead of going with
eglot. I still can switch to the "eglot" branch which is working
perfectly, but it feels is not worth the hassle. I can still configure
lsp-mode to be minimal, and benefit from the great community of
lsp-mode (eglot also has a great one).