feat: configure org-src-mode, eww brwoser and devdocs
Configure keybindings for org-src-mode, fitting the style of bindings
for temporal buffers that I have defined: "c"ancel and "f"inalize.

Update lsp-mode bindings, following a coding session I had with Java. I
seem to be using lsp-find-references a lot more than lsp-rename, moving
the former to "r", and the latter to "R".

Also hook devdocs bindings to the major-mode-definer for prog-mode.